fix firebase
