+++ src/emulator/storage/apis/firebase.ts
@@ -1,10 +1,10 @@
 import { EmulatorLogger } from "../../emulatorLogger";
 import { Emulators } from "../../types";
 import * as uuid from "uuid";
-import { gunzipSync } from "zlib";
 import { IncomingMetadata, OutgoingFirebaseMetadata, StoredFileMetadata } from "../metadata";
 import { Request, Response, Router } from "express";
 import { StorageEmulator } from "../index";
+import { sendFileBytes } from "./shared";
 import { EmulatorRegistry } from "../../registry";
 import { parseObjectUploadMultipartRequest } from "../multipart";
 import { NotFoundError, ForbiddenError } from "../errors";
@@ -121,44 +121,7 @@ export function createFirebaseEndpoints(emulator: StorageEmulator): Router {
 
     // Object data request
     if (req.query.alt === "media") {
-      let didGunzip = false;
-      if (metadata.contentEncoding === "gzip") {
-        const acceptEncoding = req.header("accept-encoding") || "";
-        const shouldGunzip = !acceptEncoding.includes("gzip");
-        if (shouldGunzip) {
-          data = gunzipSync(data);
-          didGunzip = true;
-        }
-      }
-      res.setHeader("Accept-Ranges", "bytes");
-      res.setHeader("Content-Type", metadata.contentType || "application/octet-stream");
-      res.setHeader("Content-Disposition", metadata.contentDisposition || "inline");
-      setObjectHeaders(res, metadata);
-      if (didGunzip) {
-        // Don't populate Content-Encoding if decompressed. See
-        // https://cloud.google.com/storage/docs/transcoding#decompressive_transcoding.
-        res.removeHeader("Content-Encoding");
-      }
-
-      // Content Range headers should be respected only if data was not decompressed, see
-      // https://cloud.google.com/storage/docs/transcoding#range.
-      const shouldRespectContentRange = !didGunzip;
-      if (shouldRespectContentRange) {
-        const byteRange = req.range(data.byteLength, { combine: true });
-
-        if (Array.isArray(byteRange) && byteRange.type === "bytes" && byteRange.length > 0) {
-          const range = byteRange[0];
-          res.setHeader(
-            "Content-Range",
-            `${byteRange.type} ${range.start}-${range.end}/${data.byteLength}`
-          );
-          // Byte range requests are inclusive for start and end
-          res.status(206).end(data.slice(range.start, range.end + 1));
-          return;
-        }
-      }
-      res.end(data);
-      return;
+      return sendFileBytes(metadata, data, req, res);
     }
 
     // Object metadata request
